Spread out the digit calculation over the displays & refresh time every hour
Last commit heavily front-loaded the time calculation. It got the time and determined every digit to display AND THEN displayed each of those digits. Now it gets the time and the two digits for one display, then the two digits for the next display, etc. This should prevent the glitches in the display updates.

This also reuses the hardware timer to re-sync the time from the internet every hour. All it does is set a flag so the other core will get the time so this also shouldn't affect the display.